explorer/indexer/requirements.txt
psycopg2-binary==2.9.9
redis==5.0.1
requests==2.31.0
python-dotenv==1.0.0
structlog==23.2.0
prometheus-client==0.19.0

